third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_allow_top_navigation-1-expected.txt

This is a testharness.js-based test.
FAIL Frames with `allow-top-navigation` should be able to navigate the top frame. assert_equals: expected "can navigate" but got "cannot navigate"
Harness: the test ran to completion.
